backend/requirements.txt
Django==4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2
django-allauth==0.63.6
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
djangorestframework-simplejwt==5.3.1